initial generic implementation
This is an attempt at quicktest v2.

We port a lot of the existing quicktest API but use a generic-based implementation
rather than `interface{}` throughout, which provides additional type security.

We leave out all the stuff from the old API that is no longer needed: for example:

- the `*qt.C` type - without generic methods and with no need for any state, this doesn't seem useful any more.
- `Defer` - superceded by `T.Cleanup`.
- `Setenv` - superceded by `T.Setenv`.
- `Unsetenv` - trivial to implement on top of `T.Setenv`.

The generic API is a little awkward to use in places, but harder to get wrong.
For example, one might hope that this would work:

```
x := int64(3)
qt.Assert(t, x, qt.Equals(3))
```

but it doesn't because of the way that type-propagation works (bottom up) in Go generics.
Instead, you have to do:

```
qt.Assert(t, x, qt.Equals(int64(3)))
```
instead.
But at least the compiler will tell you about that, and something like this, which was dangerously fragile in the old version, is fine now because you get a compiler error where it would previously have succeeded erroneously in the past.

```
qt.Assert(t, x, qt.Not(qt.Equals(3)))
```

I've left the reflection in place for `qt.Any` and `qt.All` (and hence `qt.Contains` too), because it was going to be too onerous to use otherwise, particularly for maps.
That's because a primitive like `qt.MapContains(5)` needs to be generic over all maps with `int` values, but that's not possible to express with Go generics.

An alternative API that retains the composible nature of quicktest might be to pass the `got` value into each checker.
For example:

```
	Assert(t, Equals(x, 99))
```

POC [here](https://go.dev/play/p/Y2ZvIB-nFx7?v=gotip). After thinking about that for a bit, I think it's probably preferable, so we should probably abandon this PR, but it'll be a few more days of work, so I'm leaving this for the moment.