revised generic implementation
This is a second attempt at quicktest v2, revised to avoid the awkward lack
of type propagation between "got" and "want" values.

We port a lot of the existing quicktest API but use a generic-based implementation
rather than `interface{}` throughout, which provides additional type security.

We leave out all the stuff from the old API that is no longer needed: for example:

- the `*qt.C` type - without generic methods and with no need for any state, this doesn't seem useful any more.
- `Defer` - superceded by `T.Cleanup`.
- `Setenv` - superceded by `T.Setenv`.
- `Unsetenv` - trivial to implement on top of `T.Setenv`.

The generic API is a little more verbose to use than previously, but harder to get wrong.
Where before, we'd have done:

```
c.Assert(a, qt.Equals, 42)
```

in the new implementation this is 4 characters longer:

```
qt.Assert(t, qt.Equals(a, 42))
```

As an example of how the new API is safer, consider this code using the old API:

```
var x int64
qt.Assert(t, x, qt.Not(qt.Equals(0)))		// Note: succeeds wrongly!
```

The above check would always have succeeded because `int64` and `0` are different types.

With the new API, this will fail appropriately

```
var x int64
qt.Assert(t, x, qt.Not(qt.Equals(0)))
```

The basic paradigm is that a `qt.Checker` is a kind of symbolic representation
of the expression to be checked which allows the quicktest code to introspect
the arguments in order to print them out nicely.

For better type safety, I've introduced new slice- and map-specific container operations
rather than relying on reflection, so you'll need to use `qt.SliceContains` or `qt.StringContains` instead
of just `qt.Contains`. This is a bit of a trade-off - we could still potentially use reflection
if we thought that the reduced verbosity was worth it.