Call TB.Helper from Assert and Check
Assertions and Checks currently don't call TB.Helper() directly, which
leads to test output like the following:

    === FAIL: . TestMapBatch/PerCPUArray (0.01s)
        quicktest.go:12:
            error:
              got non-nil value
            got:
              e"batch update: not supported"
            stack:
              /home/runner/actions-runner/_work/ebpf/ebpf/map_test.go:148
                qt.Assert(t, qt.IsNil(err))

Note that the first source location is that of qt.Assert. The correct
location can still be found from the stack trace, but that is at the end
of the message instead of at the start.

Invoke TB.Helper in Assert and Check so that the Go testing package
reports a more accurate / interesting location.